video/requirements.txt
SQLAlchemy==2.0.2
grpcio==1.51.1
grpcio-tools==1.51.1
sqlparse==0.4.2
SQLAlchemy-Utils==0.40.0